+++ .swiftlint.yml
@@ -368,7 +368,7 @@ only_rules:
   - yoda_condition
 
 attributes:
-  attributes_with_arguments_always_on_line_above: false
+  attributes_with_arguments_always_on_line_above: true
 
 deployment_target: # Availability checks or attributes shouldn’t be using older versions that are satisfied by the deployment target.
   iOSApplicationExtension_deployment_target: 16.0
