+++ Sources/SpeziStudy/Study Manager/StudyManager+Other.swift
@@ -1,0 +1,78 @@
+//
+// This source file is part of the Stanford Spezi open-source project
+//
+// SPDX-FileCopyrightText: 2025 Stanford University and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+
+import Foundation
+import SpeziScheduler
+import SwiftUI
+
+
+extension Task.Context {
+    // ISSUE for some reason only .json works? .propertyList (the default) fails to decode the input?!
+    
+    /// The ``Action`` associated with the task.
+    @Property(coding: .json)
+    public var studyScheduledTaskAction: StudyManager.ActionCard.Action?
+}
+
+
+extension Task.Category {
+    /// An informational task, e.g. an article the user should read
+    public static let informational = Self.custom("edu.stanford.spezi.SpeziStudy.task.informational")
+}
+
+
+extension View {
+    /// Configures the SpeziStudy-specific task category appearances
+    public func injectingCustomTaskCategoryAppearances() -> some View {
+        self.taskCategoryAppearance(for: .informational, label: "Informational", image: .system("text.rectangle.page"))
+    }
+}
+
+
+extension SpeziScheduler.Schedule {
+    /// An Error that can occur when attempting to create a `SpeziScheduler.Schedule` from a `StudyDefinition.ScheduleElement`.
+    public enum ScheduleConversionError: Error {
+        case unsupportedInput
+    }
+    
+    /// - parameter other: the study definition schedule element which should be turned into a `Schedule`
+    /// - parameter participationStartDate: the date at which the user started to participate in the study.
+    init(_ other: StudyDefinition.ScheduleElement, participationStartDate: Date) throws(ScheduleConversionError) {
+        switch other.componentSchedule {
+        case .once(.studyBegin, let offset):
+            self = .once(at: participationStartDate.advanced(by: offset.totalSeconds))
+        case .once(.studyEnd, _):
+            // instead of throwing an error here, we probably wanna return nil,
+            // since this is something where we do want to continue processing the remaining schedule elements!!!
+            throw .unsupportedInput
+        case .once(.completion, offset: _):
+            throw .unsupportedInput
+        case let .repeated(.daily(interval, hour, minute), startOffsetInDays):
+            self = .daily(
+                interval: interval,
+                hour: hour,
+                minute: minute,
+                second: 0,
+                startingAt: participationStartDate.addingTimeInterval(60 * 60 * 24 * TimeInterval(startOffsetInDays)),
+                end: .never,
+                duration: .tillEndOfDay
+            )
+        case let .repeated(.weekly(interval, weekday, hour, minute), startOffsetInDays):
+            self = .weekly(
+                interval: interval,
+                weekday: weekday,
+                hour: hour,
+                minute: minute,
+                second: 0,
+                startingAt: participationStartDate.addingTimeInterval(60 * 60 * 24 * TimeInterval(startOffsetInDays)),
+                end: .never,
+                duration: .tillEndOfDay
+            )
+        }
+    }
+}
