+++ Tests/UITests/TestApp/TestApp.swift
@@ -6,8 +6,8 @@
 // SPDX-License-Identifier: MIT
 //
 
-import SwiftUI
 import SpeziStudy
+import SwiftUI
 
 
 @main
